==> pre-process/03-doc_embedding_projection.ipynb <==
# Copyright (c) 2022, salesforce.com, inc.
# All rights reserved.
# SPDX-License-Identifier: BSD-3-Clause
# For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause

This notebook can be run directly on Google Colab. 

Notice that you need to specify the directory/filename to save the document projection vectors.

To read or write files on Google Drive, you can run the code below to mount your Google Drive in the notebook:

```
from google.colab import drive
drive.mount('/content/drive')
```
Then you can specify the path to the file on the Google drive by a path string starting with "/content/drive/MyDrive/"

# !pip install datasets

'''
  Load data.
  You can either download the data again, or just reload the data which has been
  saved as doc.jsonl from 02-doc_process (train+test)
'''
doc_list = []

'''option 1: download the mnli and reconstruct the input'''
# from datasets import load_dataset
# dataset = load_dataset('multi_nli', split='validation_matched')
# dataset = dataset.filter(lambda x: x['genre']=='travel')
# for item in dataset:
#     doc_list.append(item['premise']+'<S/>'+item['hypothesis'])

'''option 2: read the json file and reconstruct the input'''
import pandas as pd
df = pd.read_json(path_or_buf="<specify your path>/mnli_government_travel/doc.jsonl", orient="records")
for idx, item in df.iterrows():
    doc_list.append(item['sentence1']+'<S/>'+item['sentence2'])

doc_list[:2]

Generate Document Embedding

!pip install -U sentence-transformers

from sentence_transformers import SentenceTransformer
model = SentenceTransformer("paraphrase-multilingual-mpnet-base-v2")

sentence_embeddings = model.encode(doc_list)

sentence_embeddings.shape

## Dimensionality Reduction 
run t-SNE algorithm to map the document embedding to 2D space.

from sklearn.manifold import TSNE

X_embedded = TSNE(n_components=2).fit_transform(sentence_embeddings)

np.savetxt("<spcify your path>/mnli_government_travel/sentence_tsne.csv", X_embedded, delimiter=",")
