==> pre-process/04-high_level_feat.ipynb <==
# Copyright (c) 2022, salesforce.com, inc.
# All rights reserved.
# SPDX-License-Identifier: BSD-3-Clause
# For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause

import pandas as pd
import numpy as np

import nltk
import spacy

import os
import json

import altair as alt

df = pd.read_json("./mnli_government_travel/doc.jsonl", lines=False, orient='records')

df.head()

df.shape

model_output = pd.read_csv(filepath_or_buffer = "./mnli_government_travel/model_output.csv", header = 0)
model_output

df['y_pred'] = model_output['y_pred'].values
df['y_gt'] = model_output['y_gt'].values

df.head()

is_error = (df['y_gt'] != df['y_pred']).astype(int)
is_error

is_error.sum()/is_error.shape[0]

key_list = ['y_gt', 'y_pred']

to_save = {}
charts = []
for key in key_list:
    stat_df = pd.DataFrame()
    stat_df[key] = df[key]
    stat_df['is_error'] = is_error
    to_render = stat_df.groupby([key]).sum().reset_index()
    to_render['tot'] = stat_df.groupby([key]).count().reset_index()['is_error']
    charts.append(alt.Chart(to_render).mark_bar(opacity=.5).encode(
        x= key+':N',
        y= 'tot:Q'
    )+alt.Chart(to_render).mark_bar().encode(
        x= key+':N', 
        y= 'is_error:Q'
    ))
    
    to_save['by_'+key] = to_render.to_dict("index")

alt.hconcat(charts[0],charts[1]).resolve_scale(
    y='shared'
)

'''save to file'''
with open("./mnli_government_travel/model_stat.json", "w") as json_output:
    json_output.write(json.dumps(to_save))

## Extract high-level features:  percentage of noun/adj/adv/verb/num/pron.

'''tokenization'''
from gensim.utils import simple_preprocess

data_word_list = [simple_preprocess(x['sentence1']+" <S> "+x['sentence2']) for i,x in df.iterrows()]

premise = [x['sentence1'] for i,x in df.iterrows()]
hypothesis = [x['sentence2'] for i,x in df.iterrows()]

nlp = spacy.load('en_core_web_sm')
pssage_ready = []

for sent in data_word_list:
    # Parse the sentence using the loaded 'en' model object `nlp`. Extract the lemma for each token and join
    doc = nlp(" ".join(sent))
    pssage_ready.append([token.pos_ for token in doc])
    

p_count = np.unique(pd.DataFrame(pssage_ready)[0], return_counts=True)

allowed_postags=['ADJ', 'ADV', 'NOUN', 'NUM', 'VERB', 'PRON']


np.unique(pssage_ready[1], return_counts=True) 

word_type_dict = {}
word_type_dlist = []

for i in range(len(pssage_ready)):
    p_count = np.unique(pssage_ready[i], return_counts=True)
    word_type_dict = {}
    for ix, word in enumerate(p_count[0]):
        if word in allowed_postags:
            word_type_dict[word] = p_count[1][ix]/len(pssage_ready[i])
    word_type_dlist.append(word_type_dict)  

pd.DataFrame(word_type_dlist)

doc_pos_df = pd.DataFrame(word_type_dlist).fillna(0)
doc_pos_df.head()

## Extract high-level features: document length, overlap extent

from collections import Counter

def get_overlapping_rate(list1, list2):
    overlap = len(list((Counter(list1) & Counter(list2)).elements()))
    tot = np.min([len(list1), len(list2)])
    return overlap/tot

overlap_rate = []
doc_len = []
for i in range(len(data_word_list)):
    doc_len.append(len(data_word_list[i]))
    overlap_rate.append(get_overlapping_rate(premise[i], hypothesis[i]))


hfeat_df = pd.DataFrame(word_type_dlist).fillna(0)
hfeat_df['doc_len'] = doc_len
hfeat_df['overlap'] = overlap_rate
hfeat_df['pred'] = df['y_pred']

# use either pred or ground truth for rule generation
# hfeat_df['label'] = df['y_gt']

hfeat_df.head()

hfeat_df.columns

## Rule Extraction
features:
- doc length
- q/a overlapping
- pos tagging percentage
- labels
- model predictions

features may be included for other tasks/models/data:
- qtype

is_error.values

np.unique(is_error, return_counts=True)

def output_rules(to_output, columns, thresholds, dataname):
    filename = "./" + dataname + "/list.json"
    directory = os.path.dirname(filename)
    if not os.path.exists(directory):
        os.makedirs(directory)
    
    with open(filename, 'w') as output:
        output.write(json.dumps(to_output))
        
    filename = "./" + dataname + "/test.json"
    '''get column frequency'''

    with open(filename, 'w') as output:
        output.write(json.dumps({
            'columns': columns,
            'thresholds': thresholds,
        }))

import debug_rule

filter_threshold = {
    'support': 20,
    'err_rate': .27,
}

drule_obj = debug_rule.DebugRule()

drule_obj.initialize(hfeat_df.values, is_error.values, filter_threshold, verbose=True)\
.numerical2ordinal().extract_high_level_rule()

hfeat_df.values.shape

drule_obj.calculate_pval()
drule_obj.calculate_ci()

def generate_histogram(rule_lists):
    num_bin = 20
        
    hist = np.zeros(num_bin)
    for rule in rule_lists:
        pos_bin = int(np.floor(rule['err_rate'] * num_bin))
        if (pos_bin==num_bin):
            pos_bin = num_bin-1
        hist[pos_bin] += 1
    return hist

to_output = {'rule_lists': drule_obj.rules, 'target_names': ['correct', 'errors'], "top_list": drule_obj.top_hfeat_list}
hist = generate_histogram(drule_obj.rules)

to_output['histogram'] = hist.tolist()

output_rules(to_output, hfeat_df.columns.tolist(), drule_obj.thresholds.tolist(), 'mnli_government_travel_hfeat')

ordinal_df = pd.DataFrame(data=drule_obj.all, columns=hfeat_df.columns)

# add y_label (ground truth)
ordinal_df['label'] = model_output['y_gt']

# save to disk
ordinal_df.to_csv(path_or_buf="./mnli_government_travel/hfeat_stat.csv", index=None)
